requirements.txt
aiohttp
aiosqlite
hikari
hikari-tanchi
hikari-tanjun
hikari-miru
python-dotenv
uvloop; sys_platform != "win32"
